+++ src/main/java/com/crm/web/api/customer/CustomerResponse.java
@@ -7,9 +7,9 @@ import lombok.Data;
 @Data
 public class CustomerResponse {
 
-    private final Long id;
+    private final Integer id;
     private final String name;
     private final String surname;
-    private final Long createdByUserId;
-    private final Long lastModifiedByUserId;
+    private final Integer createdByUserId;
+    private final Integer lastModifiedByUserId;
 }
